+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Polyverse Boost Automatic Application Modernization
 - Use .boostignore and .gitignore files only for source exclusion - and stop using .vscodeignore (which is specific to Extension development)
 - Reduce redundant status updates and flicker
 - Source-level problems include the display category (e.g. Compliance, Security, etc.) for easier filtering
+- Enable UI-based refresh of Analysis Summary for a single file
 
 ### Bug Fixes
 - Fixed issue preventing output generation when Notebook conversion language or test framework is changed via UI toolbar